Read more about Shop Manager here
THEMATICAREA:                  COMMERCE
PROBLEM:	 INEFFICIENT RETAILING 
PROPOSED SOLUTION:	SHOP MANAGER 
 
  Who is affected by the problem or challenge and how they suffer?
Shop keepers are mostly affected by the problem.
In Uganda, Business men and women have continuously faced the problem in allocating products in their shops especially for big shops hence keeping a customer on waiting. The sometimes claim that a product is out of stock yet just allocating was the problem.
Some of them sometimes sell expired products but unknowingly.
Some businesses run out of stock of some products but can’t identify it until when a customer comes to buy a product and to find, it is no where to be seen.
In cases of introduction of new workers, it takes a wrong process and period for one to get in mind the prices of all products in stock.
According to WHO, at least US$550,000 worth of antiretrovirals and 10 million antimalarial doses recently expired in National Medical Stores (NMS).
According to the delivery report of 25thFeb-2ndMarch, 2020 by the National Medical Stores, Government is to collect and dispose off expired and unwanted medicines from health facilities around the Country.
The Uganda Consumer’s Protection Association (UCPA) has said major supermarket and select shops in the city are among the outlets that sell food products that are expired (New Vision, Kampala 30thMarch, 2020).
 
Proposed approach for a solution to this problem.
Diaz Business Developer shall have a function that divides the shop into nine parts for easy allocation of a product. This feature requires the user to enter the location where he/she has placed the product when adding it to stock and will display the same location in case the product has been requested for sale.
DBD will come with a function that notifies the user/seller of the expiry date every after 10hrs when it comes below Five days to avoid selling of expired products and keep the shop owner aware of yet to expire products.
DBD will notify the seller of the products running out of stock every after 10hrs when the amount in stock comes below five units to enable him enable him update stock on time.
In an attempt to sell a product, DBD will notify the seller of the price of sale for the product hence reducing the burden of having the keep in mind the prices of all products in stock.

How ICT’s and networks would be used in the implementation of your solution?
The application shall have inputs like input forms which shall be used to input data about products into a repository (central database) in the repository, the data collected shall be processed using machine leaning and artificial intelligence algorithms like decision tree algorithms and other decision support systems. The user (shopkeeper) shall access the input data from the database at any time.
The application shall bear a link to a website which shall act as the user manual with all the relevant information about the application and its usage.

How the affected people might need to change their behavior to use your solution?
Users like Shopkeepers, hospital workers shall need to be trained on the application usage.
The government needs to embrace and develop relevant policies that will support the use of the current technologies like the DBD one to solve the problem of inefficient retailing processes.

The impact of the solution on Uganda’s economy?
If the application is implemented, there will be reduction of cases expired products in the market hence reducing losses of capital out of dumping. This will help to improve on the country’s GDP
There shall also be reduction of time spent in allocation of commodities in shops hence diverting the previous wasted time to other activities. This will improve on the level of production in these shops as well as profitability hence promoting growth of Uganda’s economy.
 
APPENDICES
References.
1.	New Vision (Kampala), 30thMarch,2020
2.	National Medical Stores delivery report of 25thFeb-2ndMarch.